requirements.txt
absl-py==0.13.0
alembic==1.4.1
altair==4.1.0
anaconda-client==1.9.0
anaconda-navigator==2.0.3
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.2.1
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
atom-ml==4.5.0
attrs==21.2.0
audioread==2.1.9
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.tempfile==1.0
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
base58==2.1.0
beautifulsoup4==4.10.0
binaryornot==0.4.4
bleach==4.0.0
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
category-encoders==2.2.2
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
clyent==1.2.2
colorama==0.4.4
conda==4.10.3
conda-build==3.21.4
conda-content-trust==0+unknown
conda-package-handling==1.7.3
conda-repo-cli==1.0.4
conda-token==0.3.0
conda-verify==3.4.2
cookiecutter==1.7.3
cryptography==35.0.0
cycler==0.10.0
databricks-cli==0.14.3
deap==1.3.1
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
depthai==2.5.0.0
dill==0.3.3
docker==5.0.0
entrypoints==0.3
featuretools==0.24.1
filelock==3.3.1
Flask==1.1.2
Flask-BasicAuth==0.2.0
flatbuffers==1.12
fonttools==4.25.0
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.17
glob2==0.7
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
gplearn==0.4.1
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.2.1
htmlmin==0.1.12
idna==3.2
ImageHash==4.2.0
imbalanced-learn==0.8.0
importlib-metadata==4.8.1
ipykernel==6.4.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.8.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libarchive-c==2.9
librosa==0.8.1
lightgbm==3.2.1
llvmlite==0.37.0
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
menuinst==1.4.16
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mlflow==1.17.0
multimethod==1.4
munkres==1.1.4
navigator-updater==0.2.1
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numba==0.54.1
numexpr==2.7.3
numpy==1.20.3
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pandas-profiling==3.0.0
pandocfilters==1.4.3
parso==0.8.2
phik==0.11.2
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.2
pkginfo==1.7.1
pooch==1.4.0
poyo==0.5.0
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
prompt-toolkit==3.0.20
protobuf==3.17.2
psutil==5.8.0
pyaml==20.4.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
pydantic==1.8.2
pydeck==0.7.0
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==3.0.4
pyreadline==2.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==1.0.4
python-slugify==5.0.2
pytz==2021.3
pywin32==228
pywinpty==0.5.7
PyYAML==6.0
pyzmq==22.2.1
QtPy==1.10.0
querystring-parser==1.2.4
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
ruamel-yaml-conda==0.15.100
scikeras==0.3.3
scikit-learn==1.0.1
scikit-optimize==0.8.1
scipy==1.7.1
Send2Trash==1.8.0
setuptools==58.0.4
shap==0.39.0
sip==4.19.13
six==1.16.0
slicer==0.0.7
smmap==4.0.0
sniffio==1.2.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
sqlparse==0.4.1
stopit==1.1.2
streamlit==0.87.0
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==2.2.0
tornado==6.1
TPOT==0.11.7
tqdm==4.62.3
traitlets==5.1.0
typeguard==2.12.1
tzdata==2021.1
tzlocal==3.0
update-checker==0.18.0
urllib3==1.26.7
validators==0.18.2
visions==0.7.1
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
Werkzeug==2.0.2
wheel==0.37.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.12.1
xeno-canto==2.0.1
xgboost==1.4.2
xmltodict==0.12.0
zipp==3.6.0
